+++ src/features/chromatin-models/components/Plot.jsx
@@ -8,7 +8,7 @@ import { jsPDF } from 'jspdf';
 import { useControls, button, levaStore } from 'leva';
 import { SVGRenderer } from 'three/examples/jsm/renderers/SVGRenderer';
 import { svg2pdf } from 'svg2pdf.js';
-import { max } from 'd3';
+import { generateRainbowColors } from '../../../utils/displayUtils';
 const Plot = () => {
   //index of the points that are clicked
   const traceCtx = useContext(TraceContext);
@@ -28,6 +28,9 @@ const Plot = () => {
   const isPerimeter = traceCtx.mode === '3' ? true : false;
   const isRadius = traceCtx.mode === '1' ? true : false;
   const isLinkage = traceCtx.mode === '2' ? true : false;
+
+  //color scale
+  const rainbowColors = generateRainbowColors(data.length);
 
   //linkage
   const clickedHandler = traceCtx.clickedHandler;
@@ -214,7 +217,7 @@ const Plot = () => {
   const colorPoint = (point) => {
     if (isPerimeter) {
       if (pointX === -1 && pointY === -1 && pointZ === -1) {
-        return 'black';
+        return rainbowColors[point];
       }
       //color when point is clicked
       if (pointX === point || pointY === point || pointZ === point) {
@@ -224,7 +227,7 @@ const Plot = () => {
       return 'white';
     }else if(isRadius){  
       //check if the point is within the radius
-      if(current === -1) return 'black';
+      if(current === -1) return rainbowColors[point];
       if(point === current) return color;
       if (points[current].distanceTo(points[point]) < radius) {
         return 'black';
@@ -235,7 +238,7 @@ const Plot = () => {
     } else {
       //default color
       if (pointA === -1 && pointB === -1) {
-        return 'black';
+        return rainbowColors[point];
       }
       //color when point is clicked
       if (pointA === point || pointB === point) {
