+++ src/utils/displayUtils.js
@@ -1,3 +1,5 @@
+
+import * as d3 from 'd3';
 /**
 * 
 * @param data [{readout:number,pos:{x:number,y:number,z:number}}}}] 
@@ -144,3 +146,22 @@ export function extractFields(dataObject) {
     chromosome: dataObject.chromosome || null,
   };
 }
+
+export const generateRainbowColors = (numSegments) => {
+  const colorStops = [
+    '#FF0000', // Red
+    '#FF7F00', // Orange
+    '#FFFF00', // Yellow
+    '#00FF00', // Green
+    '#0000FF', // Blue
+    '#4B0082', // Indigo
+    '#8B00FF'  // Violet
+  ];
+
+  const rainbowScale = d3.scaleLinear()
+    .domain(colorStops.map((_, i) => i * (numSegments - 1) / (colorStops.length - 1)))
+    .range(colorStops)
+    .interpolate(d3.interpolateRgb);
+
+  return d3.range(numSegments).map(i => rainbowScale(i));
+};
